use semver.Compare for version check
